Graficos e botões para automação

+++ app/src/main/java/br/senac/sp/aplicativopiv2/UserRecycler.kt
@@ -1,5 +1,5 @@
 package br.senac.sp.aplicativopiv2
 
-data class UserRecycler (val date: String, val potencia: Double, val gasto: Int) {
+data class UserRecycler (val date: String, val potencia: Double, val gasto: Double) {
 
 }
